fix(#48) local build of @callstack/out-of-tree-platforms
