+++ packages/rn-tester/package.json
@@ -25,7 +25,8 @@
   "dependencies": {
     "flow-enums-runtime": "^0.0.6",
     "invariant": "^2.2.4",
-    "nullthrows": "^1.1.1"
+    "nullthrows": "^1.1.1",
+    "@callstack/out-of-tree-platforms": "0.74.0"
   },
   "peerDependencies": {
     "react": "18.2.0",
